tests/wpt/tests/pointerevents/pointerevent_attributes_nohover_pointers.html

<!doctype html>
<html>
<head>
  <title>Pointer Events properties tests</title>
  <meta name="viewport" content="width=device-width">
  <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
  <style>
    html {
      touch-action: none;
    }

    div {
      padding: 0;
    }

    #square1 {
      background-color: green;
      border: 1px solid black;
      height: 50px;
      width: 50px;
      margin-bottom: 3px;
      display: inline-block;
    }

    #innerFrame {
      position: relative;
      margin-bottom: 3px;
      margin-left: 0;
      top: 0;
      left: 0;
    }
  </style>
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<!-- Additional helper script for common checks across event types -->
<script type="text/javascript" src="pointerevent_support.js"></script>
<script>

  window.onload = async () => {
     const event_list = [
      'pointerover',
      'pointerenter',
      'pointerdown',
      'pointerup',
      'pointerout',
      'pointerleave',
      'pointermove'
    ];

    function checkPointerEventAttributes(testPrefix, event, expectations) {
      const pointerTestName =
          `${testPrefix} touch.${expectations.type}`;

      test(function() {
        assert_equals(event.type, expectations.type);
      }, `${pointerTestName}.type should be ${expectations.type}`);

      test(function() {
        assert_equals(event.button, expectations.button);
      }, `${pointerTestName}.button should be ${expectations.button}`);

      test(function() {
        assert_equals(event.buttons, expectations.buttons);
      }, `${pointerTestName}.buttons should be ${expectations.buttons}`);

      // Bounding rect of the event target must contain (clienX, clientY).
      const boundingRect = event.target.getBoundingClientRect();
      test(function() {
        assert_true(
            boundingRect.left <= event.clientX &&
            boundingRect.right >= event.clientX);
      }, `${pointerTestName}.clientX is within the expected range`);
      test(function() {
        assert_true(
            boundingRect.top <= event.clientY &&
            boundingRect.bottom >= event.clientY);
      }, `${pointerTestName}.clientY is within the expected range`);

      check_PointerEvent(event, testPrefix);

      // Test isPrimary
      test(function () {
        assert_equals(event.isPrimary, true);
      }, `${pointerTestName}: isPrimary attribute is true.`);
    }

    function injectScrub(element) {
      return new test_driver.Actions()
        .addPointer('pointer1', 'touch')
        .pointerMove(0, -20, {origin: element})
        .pointerDown()
        .addTick()
        .addTick()
        .pointerMove(0, 20, {origin: element})
        .addTick()
        .addTick()
        .pointerUp()
        .send();
    }

    async function tapDone() {
      const done_button = document.getElementById('done');
      const pointerupPromise = getEvent('pointerup', done_button);
      const actionPromise = new test_driver.Actions()
        .addPointer('pointer1', 'touch')
        .pointerMove(0, 0, {origin: done_button})
        .pointerDown()
        .addTick()
        .addTick()
        .pointerUp()
        .send();
      return actionPromise.then(pointerupPromise);
    }

    const test_fixture = async_test("All events handled");
    const listeners = {};
    const attachListener = (testPrefix, target, type, expectations,
                            elements) => {
      expectations.type = type;
      const pointer_ids = {};
      const key = `${testPrefix} ${type}`;
      const listener = (event) => {
        if (pointer_ids[testPrefix] == undefined) {
          pointer_ids[testPrefix] == event.pointerId;
        } else {
          test(() => {
            assert_equals(event.pointerId, pointer_ids[testPrefix]);
          }, `${testPrefix} touch.pointerId matches expectation`);
        }
        // Don't let the browser handle the event to help guard against
        // potential memory leaks.
        event.preventDefault();
        checkPointerEventAttributes(testPrefix, event, expectations);
        target.removeEventListener(type, listener);
        delete listeners[key];
      };
      target.addEventListener(type, listener);
      listeners[key] = listener;
    };

    const square1 = document.getElementById("square1");
    const innerFrame = document.getElementById('innerFrame');
    const square2 = innerFrame.contentDocument.getElementById('square2');
    const expectations = {
      pointerover:  { button:  0, buttons: 1 },
      pointerenter: { button:  0, buttons: 1 },
      pointerdown:  { button:  0, buttons: 1 },
      pointermove:  { button: -1, buttons: 1 },
      pointerup:    { button:  0, buttons: 0 },
      pointerout:   { button:  0, buttons: 0 },
      pointerleave: { button:  0, buttons: 0 },
    };
    event_list.forEach(type => {
      attachListener('', square1, type, expectations[type]);
      attachListener('inner frame', square2, type, expectations[type]);
    });
    await injectScrub(square1);
    await injectScrub(square2);
    await tapDone();

    test_fixture.step(() => {
      assert_equals(
          Object.keys(listeners).length, 0,
          `Missing tests for ${Object.keys(listeners).join(', ')}`);
      test_fixture.done();
    });
  };
</script>
<body>
  <div id="square1"></div>
  <div>
    <iframe id="innerFrame" srcdoc='
      <style>
        html {
          touch-action: none;
        }
        #square2 {
          background-color: green;
          border: 1px solid black;
          height: 50px;
          width: 50px;
          display: inline-block;
        }
      </style>
      <body>
        <div id="square2"></div>
      </body>
    '></iframe>
  </div>
  <!-- Used to detect a sentinel event. Once triggered, all other events must
       have been processed. -->
  <div>
    <button id="done">done</button>
  </div>
</body>
</html>
